+++ Code/cyclegan_website/components/ImageInput.tsx
@@ -6,40 +6,12 @@ import {
   createInferenceSession,
   drawOnnxPrediction,
 } from "../utils/drawOnnxPrediction";
+import { drawTfjsPrediction, getTfjsModel } from "../utils/drawTfjsPrediction";
 
 interface ImageInputProps {
   type: string;
   modelURL: string;
   format: string;
-}
-
-async function drawTfjsPrediction(
-  model: tf.GraphModel,
-  canvas: HTMLCanvasElement,
-  image: HTMLImageElement
-) {
-  const input = tf.browser
-    .fromPixels(image, 3)
-    .toFloat()
-    .mul(1 / 127.5)
-    .sub(1)
-    .resizeBilinear([256, 256]);
-
-  try {
-    console.log("Predicting output...");
-    console.log(`input = ${input}`);
-    const output: tf.Tensor<tf.Rank> = model.predict(
-      tf.expandDims(input, 0)
-    ) as tf.Tensor<tf.Rank>;
-    console.log(`output = ${output}`);
-    const outputTensor = output.mul(127.5).add(127.5).toInt();
-    const squeezedOutput = tf.squeeze(outputTensor, [0]).as3D(256, 256, 3);
-    const context = canvas.getContext("2d");
-    context.clearRect(0, 0, canvas.width, canvas.height);
-    tf.browser.toPixels(squeezedOutput, canvas);
-  } catch {
-    console.error("Model prediction failed.");
-  }
 }
 
 const ImageInput = ({ type, modelURL, format }: ImageInputProps) => {
@@ -48,13 +20,8 @@ const ImageInput = ({ type, modelURL, format }: ImageInputProps) => {
   let inferenceSession: ort.InferenceSession;
   const sessionOptions = { executionProviders: ["wasm"] };
 
-  const getTfjsModel = async () => {
-    console.log("Loading TensorFlow.js model...");
-    return await tf.loadGraphModel(modelURL);
-  };
-
-  if (format == "tfjs") {
-    getTfjsModel().then((model: tf.GraphModel) => {
+  if (format == "tfjs" && tfModel == null) {
+    getTfjsModel(modelURL).then((model: tf.GraphModel) => {
       tfModel = model;
       console.log("Done loading TensorFlow.js model.");
     });
@@ -63,6 +30,7 @@ const ImageInput = ({ type, modelURL, format }: ImageInputProps) => {
   const drawPrediction = async (reader: FileReader) => {
     const canvas = document.getElementById(type) as HTMLCanvasElement;
     canvas.style.borderRadius = `${theme.radii.lg.value}`;
+
     const image = new Image();
     image.src = reader.result.toString();
     image.style.borderRadius = `${theme.radii.lg.value}`;
@@ -90,22 +58,31 @@ const ImageInput = ({ type, modelURL, format }: ImageInputProps) => {
     };
   };
 
+  const initializeLabel = function (
+    label: HTMLElement,
+    backgroundImage: string
+  ) {
+    label.style.backgroundImage = backgroundImage;
+    label.style.backgroundSize = "256px 256px";
+    label.style.backgroundRepeat = "no-repeat";
+    label.textContent = "";
+  };
+
   const handleChange = function (event: ChangeEvent<FormElement>) {
     event.stopPropagation();
     event.preventDefault();
 
+    const label = document.getElementById(`label-file-upload-${type}`);
+
     const input = event.target as HTMLInputElement;
     const files = input.files;
-    const label = document.getElementById(`label-file-upload-${type}`);
     const file = files[0];
+
     const reader = new FileReader();
     reader.addEventListener(
       "load",
       () => {
-        label.style.backgroundImage = `url(${reader.result})`;
-        label.style.backgroundSize = "256px 256px";
-        label.style.backgroundRepeat = "no-repeat";
-        label.textContent = "";
+        initializeLabel(label, `url(${reader.result})`);
         drawPrediction(reader);
       },
       false
@@ -154,17 +131,17 @@ const ImageInput = ({ type, modelURL, format }: ImageInputProps) => {
   }) {
     event.stopPropagation();
     event.preventDefault();
-    const files = event.dataTransfer.files;
+
     const label = document.getElementById(`label-file-upload-${type}`);
+
+    const files = event.dataTransfer.files;
     const file = files[0];
+
     const reader = new FileReader();
     reader.addEventListener(
       "load",
       () => {
-        label.style.backgroundImage = `url(${reader.result})`;
-        label.style.backgroundSize = "256px 256px";
-        label.style.backgroundRepeat = "no-repeat";
-        label.textContent = "";
+        initializeLabel(label, `url(${reader.result})`);
         drawPrediction(reader);
       },
       false
@@ -172,6 +149,7 @@ const ImageInput = ({ type, modelURL, format }: ImageInputProps) => {
     if (file) {
       reader.readAsDataURL(file);
     }
+
     setLabelTransparency(1.0);
     setLabelBackgroundColor(`${theme.colors.neutralLight.value}`);
   };
